Remove override of display_question_bank_header()
As the TODO comment states, this is a fix for Moodle version 3.4 and
earlier. Now that we're going to be supporting Moodle 4.0, I feel it's
time to drop support for this. Because the overridden method signature
has changed, so if we wanted to keep it we would have to update it.