Rebuild JavaScript with the latest Grunt version
There's a strange error occuring in some of the code generated by Grunt
when packaging. This is a hail mary to see if rebuilding with a newer
version of grunt fixes anything.

After moving the repository to the mod subdirectory in a moodle install,
I ran the following commands (because babel apparently isn't installed
and I can't be bothered to figure that out right now; Internet Explorer
is dead):

1. npm install
2. cd amd
3. grunt amd --force